add a null check
addressing:
java.lang.NullPointerException
at java.io.File.<init>(File.java:239)
at com.redhat.rhn.domain.kickstart.KickstartFactory.removeKickstartTemplatePath(KickstartFactory.java:432)
at com.redhat.rhn.domain.kickstart.KickstartFactory.removeKickstartData(KickstartFactory.java:420)
at com.redhat.rhn.manager.kickstart.KickstartDeleteCommand.store(KickstartDeleteCommand.java:54)
at com.redhat.rhn.frontend.action.kickstart.BaseKickstartEditAction.execute(BaseKickstartEditAction.java:88)

+++ java/code/src/com/redhat/rhn/domain/kickstart/KickstartFactory.java
@@ -429,10 +429,12 @@ public class KickstartFactory extends HibernateFactory {
         Profile p = Profile.lookupById(CobblerXMLRPCHelper.getAutomatedConnection(),
                 ksdataIn.getCobblerId());
         String path = getKickstartTemplatePath(ksdataIn, p);
-        File file = new File(path);
-        if (file.exists()) {
-            log.debug("deleting : " + path);
-            file.delete();
+        if (path != null) {
+            File file = new File(path);
+            if (file.exists()) {
+                log.debug("deleting : " + path);
+                file.delete();
+            }
         }
     }
